refactor(console): trace views ride the trace trigger, not span streams
The engine's devtools span streams (trace-rows / trace-spans / all-spans)
are being removed (iii-hq/iii#2088): they serialized every coalesce window
and forced this frontend to re-implement list semantics client-side
(root-wins merge, tag backfill, dropped-frame self-heal). All three
surfaces now ride the coalesced {trace_ids} tick of the trace trigger the
timeline strip already used, and re-run their own seeded, filtered
queries (notify-then-query):

- list: tick -> debounced invalidate of traces/traceGroups/
  traceGroupMembers — the append cache, tag patching and backfill
  machinery disappear, since a refetch always carries tags and cannot
  drift from the server's filter semantics
- detail: silent reload of the open trace when its id is in a tick, one
  request in flight with a trailing rerun
- masthead strip: debounced re-seed (REPLACE semantics, same read as the
  initial seed)

Pause, tab-hidden, reconnect-reseed and hover-hold behavior are
unchanged. Validated live against an engine without the streams: ticks
delivered to all three subscribers, list refetches carrying the new
traces, zero traffic while idle.
