+++ console/web/src/pages/TracesV2/hooks/useTraceData.ts
@@ -1,12 +1,7 @@
 import { useQuery, useQueryClient } from '@tanstack/react-query'
 import { useEffect, useRef, useState } from 'react'
 import { getIiiClient } from '@/lib/iii-client'
-import {
-  isAppendableTraceList,
-  mergeTraceListSpans,
-  startTraceActivityFeed,
-  startTraceListStream,
-} from '@/lib/traces-stream'
+import { startTraceActivityFeed } from '@/lib/traces-activity'
 import {
   fetchTraces,
   type TracesFilterParams,
@@ -29,15 +24,15 @@ const DEFAULT_TRACE_LIMIT = 500
  */
 const SEARCH_SEED_MAX_SPANS = 250
 
-/** Coalesce window for the trace-tags refresh: batches the trace ids that
- *  arrived from the rows stream / activity feed into one `trace_ids` read. */
-const TAG_REFRESH_DEBOUNCE_MS = 1000
-/** Minimum gap between activity-driven reseeds of a filtered/searched list
- *  (each one re-runs the parallel multi-MB window sweep). */
+/** Client-side debounce over activity ticks before refetching. The engine
+ *  already coalesces to ~one tick per 300ms window; this only collapses the
+ *  invalidate → POST round-trips of a burst of consecutive windows. */
+const ACTIVITY_REFETCH_DEBOUNCE_MS = 250
+/** Minimum gap between activity-driven reseeds of a filtered/searched list.
+ *  That seed sweeps parallel multi-MB windows (see the queryFn); riding the
+ *  short tick debounce alone would re-run it back-to-back under a busy
+ *  session and saturate the main thread with payload parses. */
 const FILTERED_RESEED_COOLDOWN_MS = 10_000
-/** Ids per refresh read. Overflow is dropped, not queued — a still-active
- *  trace re-enters via its next activity window. */
-const TAG_REFRESH_MAX_IDS = 100
 
 export interface TraceListItem {
   traceId: string
@@ -52,8 +47,7 @@ export interface TraceListItem {
   workers: string[]
   /** Root/row span attributes, normalized to a flat object. */
   attributes?: Record<string, unknown>
-  /** Trace-level tags merged by `engine::traces::list` (absent on
-   *  live-streamed rows until the next seed read). */
+  /** Trace-level tags merged by `engine::traces::list`. */
   traceTags?: Record<string, string>
 }
 
@@ -147,11 +141,10 @@ export function useTraceData({
           }
         : { spans, total, offset: 0, limit: SEARCH_SEED_MAX_SPANS }
     },
-    // This query is the one-time SEED read. Live updates arrive by APPEND over
-    // the engine `iii:devtools:trace-rows` stream (see the stream effect
-    // below), which merges new rows straight into this cache — no polling
-    // interval. Reconnect / tab-visible re-seed once to self-heal dropped
-    // frames; manual Refresh re-reads on demand.
+    // This query is the SEED read, re-run on demand: the `trace` trigger's
+    // coalesced `{trace_ids}` tick invalidates it (notify-then-query, see
+    // the effect below) — no polling interval. Reconnect / tab-visible
+    // re-seed once to recover anything missed while away.
     refetchInterval: false,
     staleTime: 1000,
     // The collector already ladders timeouts internally; stacking the
@@ -233,33 +226,28 @@ export function useTraceData({
     }
   }, [tracesData, hiddenKey])
 
-  // ── Live append over the engine `trace-rows` stream ──────────────────────
-  // The engine pushes new root rows as spans close. The UNFILTERED list merges
-  // them directly into the seed query's cache (pure append, no refetch); a
-  // filtered/searched list — and the group-by aggregate, which can't be
-  // appended — refetches on activity instead (the engine already coalesces to
-  // ~one push per window, so this is not a poll). Pause / tab-hidden freeze it;
-  // reconnect and tab-visible re-seed once to recover anything dropped while
-  // away. Subscribes once for the hook's lifetime (params are read via refs).
+  // ── Trigger-driven refetch (notify-then-query) ──────────────────────────
+  // The engine coalesces span activity into one `{ trace_ids }` tick per
+  // ~300ms window on the `trace` trigger; each tick re-runs the seeded,
+  // FILTERED queries — the engine stays the single owner of filter and tag
+  // semantics, an idle engine produces zero traffic, and there is no
+  // client-side append/merge to drift from the server's view. Pause /
+  // tab-hidden freeze it; reconnect and tab-visible re-seed once to recover
+  // anything missed while away. Subscribes once for the hook's lifetime.
   const qc = useQueryClient()
-  const mergeKeyRef = useRef<{
-    key: unknown[]
-    unfiltered: boolean
-    includeInternal: boolean
-  }>({
-    key: [],
-    unfiltered: false,
-    includeInternal: false,
-  })
-  mergeKeyRef.current = {
-    key: ['traces', filterParams, showSystem, debouncedSearch],
-    unfiltered: isAppendableTraceList(filterParams, debouncedSearch),
-    includeInternal: showSystem,
-  }
   const isPausedRef = useRef(isPaused)
   useEffect(() => {
     isPausedRef.current = isPaused
   }, [isPaused])
+  // Whether the CURRENT seed is the expensive search_all_spans sweep —
+  // read through a ref inside the once-per-lifetime subscription below.
+  const isSearchAllSeed =
+    filterParams.search_all_spans === true ||
+    Boolean(debouncedSearch && !filterParams.name)
+  const searchAllSeedRef = useRef(isSearchAllSeed)
+  useEffect(() => {
+    searchAllSeedRef.current = isSearchAllSeed
+  }, [isSearchAllSeed])
 
   useEffect(() => {
     let stop: (() => void) | undefined
@@ -267,181 +255,72 @@ export function useTraceData({
 
     const isHidden = () =>
       typeof document !== 'undefined' && document.visibilityState === 'hidden'
-    const reseed = () => {
-      qc.invalidateQueries({ queryKey: ['traces'] })
-      qc.invalidateQueries({ queryKey: ['traceGroups'] })
-      qc.invalidateQueries({ queryKey: ['traceGroupMembers'] })
-    }
-
-    // ── Trace-tags refresh / row backfill ─────────────────────────────────
-    // Streamed rows arrive WITHOUT `trace_tags` (only `traces::list` merges
-    // them), and a row's tags can change after it exists — the tag-bearing
-    // span (e.g. the harness turn step carrying `iii.session.name`) closes
-    // well after the root row was pushed. Both funnel into one debounced
-    // `trace_ids` read whose `trace_tags` are merged back into the cache; a
-    // filtered list can't be patched row-wise, so it refetches instead.
-    // The same read doubles as the list's self-heal: a trace whose
-    // `trace-rows` frame was dropped (broadcast lag, reconnect gap) keeps
-    // firing span activity, and if its fetched ROOT isn't in the cache yet
-    // it is inserted — without this, a missed row frame leaves the trace
-    // off the list until the next full reseed.
-    const pendingTagIds = new Set<string>()
-    let tagFlushTimer: ReturnType<typeof setTimeout> | undefined
-
-    const tagsEqual = (
-      a: Record<string, string> | undefined,
-      b: Record<string, string> | undefined,
-    ): boolean => {
-      const ea = Object.entries(a ?? {})
-      const eb = b ?? {}
-      return (
-        ea.length === Object.keys(eb).length &&
-        ea.every(([k, v]) => eb[k] === v)
-      )
-    }
-
-    const flushTagRefresh = async () => {
-      const ids = [...pendingTagIds].slice(0, TAG_REFRESH_MAX_IDS)
-      pendingTagIds.clear()
-      if (ids.length === 0 || isPausedRef.current || isHidden()) return
-      const { key, unfiltered, includeInternal } = mergeKeyRef.current
-      if (!unfiltered) {
-        // Row-wise patching is only sound for the append cache; a filtered
-        // list re-evaluates its filters (tags may move rows in/out of it).
-        qc.invalidateQueries({ queryKey: ['traces'] })
-        return
-      }
-      try {
-        const res = await fetchTraces({
-          trace_ids: ids,
-          include_internal: includeInternal,
-          limit: ids.length,
-        })
-        if (disposed) return
-        if (res.spans.length === 0) return
-        const tagsByTrace = new Map(
-          res.spans
-            .filter((s) => s.trace_tags)
-            .map((s) => [s.trace_id, s.trace_tags]),
-        )
-        qc.setQueryData<TracesResponse>(key, (old) => {
-          if (!old) return old
-          let changed = false
-          const known = new Set(old.spans.map((s) => s.trace_id))
-          const spans = old.spans.map((s) => {
-            const tags = tagsByTrace.get(s.trace_id)
-            if (!tags || tagsEqual(tags, s.trace_tags)) return s
-            changed = true
-            return { ...s, trace_tags: tags }
-          })
-          // Backfill: roots the rows stream never delivered (dropped
-          // frame) join the list here instead of waiting for a reseed.
-          const missing = res.spans.filter(
-            (s) => !s.parent_span_id && !known.has(s.trace_id),
-          )
-          if (missing.length === 0) {
-            return changed ? { ...old, spans } : old
-          }
-          const merged = mergeTraceListSpans(
-            spans,
-            missing,
-            DEFAULT_TRACE_LIMIT,
-          )
-          return { ...old, spans: merged, total: merged.length }
-        })
-      } catch {
-        // Transient read failure — the next activity window retries.
-      }
-    }
-
-    const scheduleTagRefresh = (ids: ReadonlyArray<string>) => {
-      for (const id of ids) pendingTagIds.add(id)
-      if (pendingTagIds.size === 0 || tagFlushTimer !== undefined) return
-      tagFlushTimer = setTimeout(() => {
-        tagFlushTimer = undefined
-        void flushTagRefresh()
-      }, TAG_REFRESH_DEBOUNCE_MS)
-    }
-
-    // Filtered/searched seeds are EXPENSIVE (parallel multi-MB windows —
-    // see the queryFn): under a busy session the rows stream would refetch
-    // them back-to-back and saturate the main thread with payload parses.
-    // Cool down to one reseed per window, trailing so the last burst lands.
+    // The search_all_spans seed is EXPENSIVE (parallel multi-MB windows —
+    // see the queryFn): its reseed rides a trailing cooldown instead of the
+    // short tick debounce, so a busy session cannot re-run it back-to-back.
     let reseedCooldownTimer: ReturnType<typeof setTimeout> | undefined
-    let lastFilteredReseed = 0
+    let lastTracesReseed = 0
+    const reseedTraces = () => {
+      lastTracesReseed = Date.now()
+      qc.invalidateQueries({ queryKey: ['traces'] })
+    }
     const throttledFilteredReseed = () => {
       const wait = Math.max(
         0,
-        lastFilteredReseed + FILTERED_RESEED_COOLDOWN_MS - Date.now(),
+        lastTracesReseed + FILTERED_RESEED_COOLDOWN_MS - Date.now(),
       )
       if (wait === 0) {
-        lastFilteredReseed = Date.now()
-        qc.invalidateQueries({ queryKey: ['traces'] })
+        reseedTraces()
         return
       }
       if (reseedCooldownTimer !== undefined) return
       reseedCooldownTimer = setTimeout(() => {
         reseedCooldownTimer = undefined
-        lastFilteredReseed = Date.now()
-        qc.invalidateQueries({ queryKey: ['traces'] })
+        reseedTraces()
       }, wait)
+    }
+    const refetchGroups = () => {
+      qc.invalidateQueries({ queryKey: ['traceGroups'] })
+      qc.invalidateQueries({ queryKey: ['traceGroupMembers'] })
+    }
+    const refetchAll = () => {
+      reseedTraces()
+      refetchGroups()
+    }
+
+    let refetchTimer: ReturnType<typeof setTimeout> | undefined
+    const scheduleRefetch = () => {
+      if (refetchTimer !== undefined) return
+      refetchTimer = setTimeout(() => {
+        refetchTimer = undefined
+        if (disposed || isPausedRef.current || isHidden()) return
+        if (searchAllSeedRef.current) {
+          throttledFilteredReseed()
+          refetchGroups()
+        } else {
+          refetchAll()
+        }
+      }, ACTIVITY_REFETCH_DEBOUNCE_MS)
     }
 
     void (async () => {
       const client = await getIiiClient()
       if (disposed) return
 
-      const offStream = startTraceListStream(client, (spans) => {
+      const offActivity = startTraceActivityFeed(client, () => {
         if (isPausedRef.current || isHidden()) return
-        const { key, unfiltered } = mergeKeyRef.current
-        if (unfiltered) {
-          qc.setQueryData<TracesResponse>(key, (old) => {
-            const merged = mergeTraceListSpans(
-              old?.spans ?? [],
-              spans,
-              DEFAULT_TRACE_LIMIT,
-            )
-            return {
-              spans: merged,
-              total: merged.length,
-              offset: 0,
-              limit: DEFAULT_TRACE_LIMIT,
-            }
-          })
-          // The appended rows carry no tags yet; fetch them once the trace's
-          // spans settle. (The filtered branch refetches, which re-reads tags.)
-          scheduleTagRefresh(spans.map((s) => s.trace_id))
-        } else {
-          throttledFilteredReseed()
-        }
-        // The group-by aggregate (and any expanded group's member list)
-        // can't be appended; refetch them on activity.
-        qc.invalidateQueries({ queryKey: ['traceGroups'] })
-        qc.invalidateQueries({ queryKey: ['traceGroupMembers'] })
-      })
-
-      // Span activity = a listed trace's merged tags may have just changed
-      // (late tag-bearing span, renamed session), OR a trace the rows
-      // stream never delivered is doing work. Both route through the same
-      // debounced read: known rows get their tags patched, unknown roots
-      // get backfilled (see `flushTagRefresh`). Gated on the seed having
-      // landed — before that the seed read itself covers everything.
-      const offActivity = startTraceActivityFeed(client, (traceIds) => {
-        if (isPausedRef.current || isHidden()) return
-        const cached = qc.getQueryData<TracesResponse>(mergeKeyRef.current.key)
-        if (!cached) return
-        scheduleTagRefresh(traceIds)
+        scheduleRefetch()
       })
 
       const offConn = client.addConnectionStateListener((state) => {
-        if (state === 'connected' && !isPausedRef.current) reseed()
+        if (state === 'connected' && !isPausedRef.current) refetchAll()
       })
 
       let offVisibility: (() => void) | undefined
       if (typeof document !== 'undefined') {
         const onVisible = () => {
           if (document.visibilityState === 'visible' && !isPausedRef.current) {
-            reseed()
+            refetchAll()
           }
         }
         document.addEventListener('visibilitychange', onVisible)
@@ -450,13 +329,12 @@ export function useTraceData({
       }
 
       stop = () => {
-        offStream()
         offActivity()
         offConn()
         offVisibility?.()
-        if (tagFlushTimer !== undefined) {
-          clearTimeout(tagFlushTimer)
-          tagFlushTimer = undefined
+        if (refetchTimer !== undefined) {
+          clearTimeout(refetchTimer)
+          refetchTimer = undefined
         }
         if (reseedCooldownTimer !== undefined) {
           clearTimeout(reseedCooldownTimer)
